chore: don't generate public commands file
