+++ packages/browser/src/node/commands/index.ts
@@ -11,16 +11,3 @@ export default {
   writeFile,
   sendKeys,
 }
-
-export type {
-  BufferEncoding,
-  FsOptions,
-} from './fs'
-
-export type {
-  TypePayload,
-  PressPayload,
-  DownPayload,
-  UpPayload,
-  SendKeysPayload,
-} from './keyboard'
